osmaxx-py/requirements/local.txt
# Local development dependencies go here
-r base.txt
django-extensions
Werkzeug
ipython

# also included in test.txt
coverage
flake8

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar
